requirements.txt
asgiref==3.10.0
boto3==1.40.61
botocore==1.40.61
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.3
daphne==4.2.1
dj-database-url==3.0.1
Django==5.2.7
django-allauth==65.12.1
django-heroku==0.3.1
django-storages==1.14.6
gunicorn==23.0.0
idna==3.11
jmespath==1.0.1
packaging==25.0
Pillow==12.0.0
psycopg2==2.9.10
pycparser==2.23
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
requests==2.32.5
s3transfer==0.14.0
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
whitenoise==6.11.0
